backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-filter==24.2
qrcode==7.4.2
cryptography==42.0.8
pytest==8.2.1
pytest-django==4.8.0
